Sacae UA Special Case Error Fix
Fixed following errors:
-When reloading game, for that turn, healing would not apply
-When a unit was trained, would not heal from moving on that turn

This stemmed from the game recordingstarting positions at the start of the turn. If the unit was trained then it wouldn't have existed when the start positions were recorded and therefore would give lua errors when that information was trying to be pulled. Reloading was a compromise, impossible to know the actual position at the start of the turn as it would have to store that information in the gamesave or on the computer, this way the heal will apply acting like the postion it was when it loaded was the start position instead to avoid the error